Split up types unit tests
There's a bunch of tests in this folder now, particularly for JSON and
string. Move tests to subfolders matching their source location.
