+++ test/types/bit_stream/bit_stream.cpp
