+++ test/types/json/json_accessors.cpp
@@ -1,4 +1,4 @@
-#include "test/types/json_helpers.hpp"
+#include "test/types/json/json_helpers.hpp"
 
 #include "fly/types/json/json.hpp"
